Implementation of stateful API for potentials.
This adds an implementation of a stateful API for potentials,
where rather than a potential being represented as a pure function,
it can maintain state. At present this is solely designed to
expose the old API through a caching layer. For potentials that
are expensive to compute from scratch this may be a performance
gain to use, but otherwise has relatively few advantages over the
old way of defining things, although in some cases it may be more
convenient to implement something using the streaming API.